requirements.txt
altair==4.2.0
attrs==22.1.0
backports.zoneinfo==0.2.1
blinker==1.5
Brotli==1.0.9
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
decorator==5.1.1
entrypoints==0.4
ffmpeg-python==0.2.0
filelock==3.8.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.29
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
Jinja2==3.1.2
jsonschema==4.17.0
MarkupSafe==2.1.1
more-itertools==9.0.0
mutagen==1.46.0
numpy==1.22
packaging==21.3
pandas==1.3.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
pyarrow==10.0.0
pycryptodomex==3.15.0
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytube==15.0.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
rich==12.6.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.27.0
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==1.13.0
tornado==6.2
tqdm==4.64.1
transformers==4.24.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
websockets==10.4
whisper @ git+https://github.com/openai/whisper.git@9e653bd0ea0f1e9493cb4939733e9de249493cfb
yt-dlp==2022.11.11
yt-whisper @ git+https://github.com/m1guelpf/yt-whisper.git@0190e7e3cdafb015b1325e30e80e558af7d7c57f
zipp==3.10.0